+++ src/layouts/Main/components/Footer/Footer.js
@@ -27,7 +27,7 @@ const Footer = props => {
           href="https://github.com/tip-2020-c1-grupo1/refugio-docs"
           target="_blank"
         >
-          Devias IO
+          Refugio IO
         </Link>
         . 2020
       </Typography>