+++ src/views/AnimalList/components/index.js
@@ -1,0 +1,2 @@
+export { default as AnimalCard } from './AnimalCard';
+export { default as AnimalsToolbar } from './AnimalsToolbar';