+++ src/views/Dashboard/Dashboard.js
@@ -1,6 +1,8 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { makeStyles } from '@material-ui/styles';
+import axios from "axios";
 import { Grid } from '@material-ui/core';
+import MaterialTable from 'material-table';
 
 import {
   Budget,
@@ -9,7 +11,7 @@ import {
   TotalProfit,
   LatestSales,
   UsersByDevice,
-  LatestProducts,
+  LatestAnimals,
   LatestOrders
 } from './components';
 
@@ -18,13 +20,44 @@ const useStyles = makeStyles(theme => ({
     padding: theme.spacing(4)
   }
 }));
+// material tables
 
 const Dashboard = (props) => {
   const classes = useStyles();
   console.log(props);
+
+  const url = 'http://localhost:8000/api/animals/';
+
+  useEffect(() => {
+    console.log('preparing components');
+    axios
+      .get(url)
+      .then(({ data }) => {
+        console.log(data);
+      });
+  }, []);
+
+  
   return (
     <div className={classes.root}>
       <h2> Welcome {props.user.name} </h2>
+      <div style={{ maxWidth: '100%' }}>
+        <MaterialTable
+          columns={[
+            { title: 'Adı', field: 'name' },
+            { title: 'Soyadı', field: 'surname' },
+            { title: 'Doğum Yılı', field: 'birthYear', type: 'numeric' },
+            { title: 'Doğum Yeri', field: 'birthCity', lookup: { 34: 'İstanbul', 63: 'Şanlıurfa' } }
+          ]}
+          data={[{ name: 'Mehmet', surname: 'Baran', birthYear: 1987, birthCity: 63 }, 
+          { name: 'Luciano', surname: 'Olmedo', birthYear: 1989, birthCity: 34 },
+          { name: 'Javier', surname: 'Perini', birthYear: 1990, birthCity: 63 },
+          { name: 'Nahuel', surname: 'Benitez', birthYear: 1992, birthCity: 34 }
+        ]}
+          title="Demo Title"
+        />
+      </div>
+
       <Grid
         container
         spacing={4}
@@ -90,7 +123,7 @@ const Dashboard = (props) => {
           xl={3}
           xs={12}
         >
-          <LatestProducts />
+          <LatestAnimals />
         </Grid>
         <Grid
           item
@@ -106,5 +139,4 @@ const Dashboard = (props) => {
 
   );
 };
-
 export default Dashboard;