+++ src/views/Dashboard/components/index.js
@@ -1,6 +1,6 @@
 export { default as Budget } from './Budget';
 export { default as LatestOrders } from './LatestOrders';
-export { default as LatestProducts } from './LatestProducts';
+export { default as LatestAnimals } from './LatestAnimals';
 export { default as LatestSales } from './LatestSales';
 export { default as TasksProgress } from './TasksProgress';
 export { default as TotalProfit } from './TotalProfit';